teacher_assistant/sub_agents/student_behaviour_tracking/agent.py
"""
Student Behaviour Tracking Subagent

This subagent monitors and analyzes student behavior patterns to help
teachers understand engagement levels and identify areas for improvement.
"""

from google.adk import ADK


class StudentBehaviourTrackingAgent:
    def __init__(self):
        self.adk = ADK()
        self.name = "Student Behaviour Tracking Agent"
        
    async def track_engagement(self, student_id: str, activity_data: dict):
        """Track student engagement during activities"""
        pass
        
    async def analyze_behavior_patterns(self, student_id: str, timeframe: str):
        """Analyze behavior patterns over a given timeframe"""
        pass
        
    async def generate_behavior_report(self, student_id: str):
        """Generate a comprehensive behavior report"""
        pass